fix(dotnet): serialize scanned confidence the way Rust does
The score is an f32 on both sides, but Rust's serializer widens it to
f64 before printing, so 0.85 comes out as 0.8500000238418579. The port
printed the shortest round-trip form of the float, 0.85, and 33 PDF
fixtures failed their metadata comparison on nothing but that.

Holding the field as a double while still assigning it from a float
reproduces Rust's digits exactly and leaves the value itself unchanged.

Also adds the run-in list split and the final heading-hierarchy
compaction, two upstream stages the port never had. Neither moves the
corpus — the patterns they fire on are rare — but they were missing.

PDF metadata 324 -> 351 of 388, fully matching 79 -> 100.

+++ dotnet/src/Xberg/Types/Metadata.cs
@@ -256,7 +256,13 @@ public sealed class PdfMetadata
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)] public uint? PageCount { get; set; }
 
     /// <summary>Highest per-page scanned-page confidence in the document, in [0, 1].</summary>
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)] public float? ScannedConfidence { get; set; }
+    /// <remarks>
+    /// Held as a double but always assigned from a <c>float</c>, so it serializes with the same
+    /// digits Rust's <c>f32</c> produces once serde widens it — 0.85f prints as
+    /// <c>0.8500000238418579</c>, not <c>0.85</c>. The value is a single-precision score either
+    /// way; only the printed form differs, and the two must agree byte for byte.
+    /// </remarks>
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)] public double? ScannedConfidence { get; set; }
 
     /// <summary>One-based page numbers graded at or above the configured confidence.</summary>
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)] public List<uint>? ScannedPages { get; set; }
